@@ -84,7 +84,8 @@ define(function(require) {
 		var conicSensor = entity.conicSensor;
 		if (!defined(conicSensor)) {
 			entity.addProperty('conicSensor');
-			entity.conicSensor = conicSensor = new ConicSensorGraphics();
+			conicSensor = new ConicSensorGraphics();
+			entity.conicSensor = conicSensor;
 		}
 
 		processCommonSensorProperties(conicSensor, conicSensorData, interval, sourceUri, entityCollection);
@@ -110,7 +111,8 @@ define(function(require) {
 		var customPatternSensor = entity.customPatternSensor;
 		if (!defined(customPatternSensor)) {
 			entity.addProperty('customPatternSensor');
-			entity.customPatternSensor = customPatternSensor = new CustomPatternSensorGraphics();
+			customPatternSensor = new CustomPatternSensorGraphics();
+			entity.customPatternSensor = customPatternSensor;
 		}
 
 		processCommonSensorProperties(customPatternSensor, customPatternSensorData, interval, sourceUri, entityCollection);
@@ -146,7 +148,8 @@ define(function(require) {
 		var rectangularSensor = entity.rectangularSensor;
 		if (!defined(rectangularSensor)) {
 			entity.addProperty('rectangularSensor');
-			entity.rectangularSensor = rectangularSensor = new RectangularSensorGraphics();
+			rectangularSensor = new RectangularSensorGraphics();
+			entity.rectangularSensor = rectangularSensor;
 		}
 
 		processCommonSensorProperties(rectangularSensor, rectangularSensorData, interval, sourceUri, entityCollection);
