@@ -15,7 +15,8 @@ define(function(require) {
 	function assignSpherical(index, array, clock, cone) {
 		var spherical = array[index];
 		if (!defined(spherical)) {
-			array[index] = spherical = new Spherical();
+			spherical = new Spherical();
+			array[index] = spherical;
 		}
 		spherical.clock = clock;
 		spherical.cone = cone;
